refactor out type transformed string util cycle
